--- a/Sources/LocalLLMCLI/Common.swift
+++ b/Sources/LocalLLMCLI/Common.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+enum Backend: String, CaseIterable {
+    case llama
+    case mlx
+    case foundationModels = "foundation-models"
+}
+
+enum LocalLLMCommandError: Error {
+    case invalidModel(String)
+}