- Improve performance of llama backend

| Name | Old | New |
|------|-----|-----|
| **avgFirstTokenLatency** | 0.1112896601 | 0.1015799840 |
| **avgPromptDecodingTime** | 0.0043340127 | 0.0055063168 |
| **avgTokensPerSecond** | 28.2394521 | 67.7651768 |
| **avgTotalTime** | 71.5933350 | 29.5615257 |
| **maxFirstTokenLatency** | 0.1292719841 | 0.1044499874 |
| **maxPromptDecodingTime** | 0.0051440001 | 0.0056489706 |
| **maxTotalTime** | 82.4180670 | 30.7829370 |
| **minFirstTokenLatency** | 0.0991740227 | 0.0972169638 |
| **minPromptDecodingTime** | 0.0037740469 | 0.0054199696 |
| **minTotalTime** | 65.5089631 | 28.0940959 |

<details>
<summary>raw</summary>

<table style="width:100%; border-collapse:collapse;">
  <tr>
    <th style="text-align:left; width:50%;">OLD</th>
    <th style="text-align:left; width:50%;">NEW</th>
  </tr>
  <tr>
    <td style="vertical-align:top; padding:8px;">
<pre><code class="language-json">[
  {
    "metrics" : [
      {
        "firstTokenLatency" : 0.1292719841003418,
        "promptDecodingTime" : 0.005144000053405762,
        "tokensPerSecond" : 29.918696983247298,
        "totalTime" : 66.85297501087189,
        "totalTokens" : 2000
      },
      {
        "firstTokenLatency" : 0.09917402267456055,
        "promptDecodingTime" : 0.0037740468978881836,
        "tokensPerSecond" : 30.531932834965623,
        "totalTime" : 65.50896310806274,
        "totalTokens" : 2000
      },
      {
        "firstTokenLatency" : 0.1054229736328125,
        "promptDecodingTime" : 0.004083991050720215,
        "tokensPerSecond" : 24.267726513177738,
        "totalTime" : 82.41806697845459,
        "totalTokens" : 2000
      }
    ],
    "name" : "Basic benchmark",
    "summary" : {
      "avgFirstTokenLatency" : 0.11128966013590495,
      "avgPromptDecodingTime" : 0.004334012667338054,
      "avgTokensPerSecond" : 28.23945211046355,
      "avgTotalTime" : 71.59333503246307,
      "maxFirstTokenLatency" : 0.1292719841003418,
      "maxPromptDecodingTime" : 0.005144000053405762,
      "maxTotalTime" : 82.41806697845459,
      "minFirstTokenLatency" : 0.09917402267456055,
      "minPromptDecodingTime" : 0.0037740468978881836,
      "minTotalTime" : 65.50896310806274
    },
    "timestamp" : "2025-09-15T08:03:53Z"
  }
]</code></pre>
    </td>
    <td style="vertical-align:top; padding:8px;">
<pre><code class="language-json">[
  {
    "metrics" : [
      {
        "firstTokenLatency" : 0.10444998741149902,
        "promptDecodingTime" : 0.005648970603942871,
        "tokensPerSecond" : 71.20365187144678,
        "totalTime" : 28.094095945358276,
        "totalTokens" : 2000
      },
      {
        "firstTokenLatency" : 0.09721696376800537,
        "promptDecodingTime" : 0.00541996955871582,
        "tokensPerSecond" : 67.10931282708418,
        "totalTime" : 29.807543992996216,
        "totalTokens" : 2000
      },
      {
        "firstTokenLatency" : 0.10307300090789795,
        "promptDecodingTime" : 0.005450010299682617,
        "tokensPerSecond" : 64.98256557293409,
        "totalTime" : 30.782937049865723,
        "totalTokens" : 2000
      }
    ],
    "name" : "Basic benchmark",
    "summary" : {
      "avgFirstTokenLatency" : 0.10157998402913411,
      "avgPromptDecodingTime" : 0.0055063168207804365,
      "avgTokensPerSecond" : 67.76517675715502,
      "avgTotalTime" : 29.56152566274007,
      "maxFirstTokenLatency" : 0.10444998741149902,
      "maxPromptDecodingTime" : 0.005648970603942871,
      "maxTotalTime" : 30.782937049865723,
      "minFirstTokenLatency" : 0.09721696376800537,
      "minPromptDecodingTime" : 0.00541996955871582,
      "minTotalTime" : 28.094095945358276
    },
    "timestamp" : "2025-09-15T07:28:09Z"
  }
]</code></pre>
    </td>
  </tr>
</table>

```
swift run -c release LocalLLMCLI benchmark --model ~/.localllmclient/huggingface/models/lmstudio-community/gemma-3-4B-it-qat-GGUF/gemma-3-4B-it-QAT-Q4_0.gguf --output ~/Desktop/results.json
```

*M1 Max*

</details>

--- 
This pull request introduces two new commands to the CLI (`benchmark` and `run`) for interacting with local LLMs, along with supporting infrastructure for backend selection and error handling. The changes add comprehensive benchmarking capabilities, support for multiple backends, and improved model downloading logic. The most important changes are grouped below:

### New CLI Commands

* Added `BenchmarkCommand` in `BenchmarkCommand.swift` to run performance benchmarks on local LLMs, supporting various modes (basic, context, batch, all), custom prompts, iterations, context/batch sizes, and automatic model downloading from HuggingFace. Results are output in JSON format with detailed metrics.
* Added `RunCommand` in `RunCommand.swift` to run a single prompt through a local LLM, supporting backend selection (`llama`, `mlx`, `foundation-models`), sampling parameters, image attachments, and model downloading logic.

### Backend and Error Handling Infrastructure

* Introduced the `Backend` enum for backend selection and the `LocalLLMCommandError` enum for consistent error handling in `Common.swift`.

### Codebase Cleanup

* Removed unnecessary imports and platform-specific code from `command.swift`, centralizing logic in the new command files.